+++ src/pages/Feed/index.js
@@ -1,38 +1,26 @@
-import React, { useState, useEffect, useCallback, useRef } from 'react'
-import { FaSearch, FaNewspaper, FaTimes } from 'react-icons/fa'
-import { MdAddAPhoto } from 'react-icons/md'
+import React, { useState, useEffect } from 'react'
+import { FaSearch } from 'react-icons/fa'
 import { useSelector } from 'react-redux'
-import ReactCrop from 'react-image-crop'
 import 'react-image-crop/dist/ReactCrop.css'
 
 import { getDashboard } from '../../services/dashboard'
-import { createPost } from '../../services/posts'
 import Header from '../../components/Header'
 import Post from '../../components/Post'
+import NewPost from '../../components/NewPost'
 
 import { 
   FeedPage,
   Content,
   LeftBar,
   RightBar,
-  NewPost,
   FeedHistory,
   OnlineFriends,
-  Friend,
-  NewPostBackground,
-  NewPostModal
+  Friend
 } from './styles'
 
 const Feed = ({ history }) => {
-  const $ = document.querySelector.bind(document)
   const [posts, setPosts] = useState([])
   const devInfo = useSelector(state => state.dev.devInfo)
-  const [upImg, setUpImg] = useState()
-  const imgRef = useRef(null)
-  const [crop, setCrop] = useState({ unit: '%', width: 30, aspect: 1 / 1 })
-  const [previewUrl, setPreviewUrl] = useState()
-  const [newPostContent, setNewPostContent] = useState('')
-  const [newPostThumbnail, setNewPostThumbnail] = useState('')
 
   useEffect(() => {
     async function callApi() {
@@ -41,93 +29,6 @@ const Feed = ({ history }) => {
     }
     callApi()
   }, [])
-
-  function openNewPostModal() {
-    document.querySelector('#newPostModalBackground').style.display = "block"
-  }
-
-  function closeNewPostModal() {
-    document.querySelector('#newPostModalBackground').style.display = "none"
-  }
-
-  async function handleNewPost(event) {
-    event.preventDefault()
-    await createPost(newPostContent, newPostThumbnail)
-    closeNewPostModal()
-    // TODO
-  }
-  async function previewImage(e) {
-    e.preventDefault()
-    const previewImg = $('.preview-img')
-    const labelPreview = $('.label-preview')
-    const cropButton = $('.cropImage')
-    const crop = $('.react-crop')
-    previewImg.src = previewUrl
-    let file = await fetch(previewUrl)
-      .then(r => r.blob())
-      .then(blobFile => new File([blobFile], "thumbnail", { type: "image/png" }))
-    setNewPostThumbnail(file)
-    labelPreview.classList.add('hidden')
-    crop.classList.add('hidden')
-    cropButton.classList.add('hidden')
-    previewImg.classList.remove('hidden')
-  }
-
-  const onSelectFile = e => {
-    if (e.target.files && e.target.files.length > 0) {
-      const [file] = e.target.files
-      const reader = new FileReader()
-      reader.addEventListener('load', () => setUpImg(reader.result))
-      reader.readAsDataURL(file)
-    }
-  }
-  
-  const onLoad = useCallback(img => {
-    imgRef.current = img
-    const labelPreview = $('.label-preview')
-    const cropImage = $('.cropImage')
-    labelPreview.classList.add('hidden')
-    cropImage.classList.remove('hidden')
-  }, [$])
-
-  const makeClientCrop = async crop => {
-    if (imgRef.current && crop.width && crop.height) {
-      createCropPreview(imgRef.current, crop, 'newFile.jpeg')
-    }
-  }
-
-  const createCropPreview = async (image, crop, fileName) => {
-    const canvas = document.createElement('canvas')
-    const scaleX = image.naturalWidth / image.width
-    const scaleY = image.naturalHeight / image.height
-    canvas.width = crop.width
-    canvas.height = crop.height
-    const ctx = canvas.getContext('2d')
-
-    ctx.drawImage(
-      image,
-      crop.x * scaleX,
-      crop.y * scaleY,
-      crop.width * scaleX,
-      crop.height * scaleY,
-      0,
-      0,
-      crop.width,
-      crop.height
-    )
-
-    return new Promise((resolve, reject) => {
-      canvas.toBlob(blob => {
-        if (!blob) {
-          reject(new Error('Canvas is empty'))
-          return
-        }
-        blob.name = fileName
-        window.URL.revokeObjectURL(previewUrl)
-        setPreviewUrl(window.URL.createObjectURL(blob))
-      }, 'image/jpeg')
-    })
-  }
 
   return (
     <FeedPage>
@@ -158,48 +59,7 @@ const Feed = ({ history }) => {
           </OnlineFriends>
         </LeftBar>
         <RightBar>
-          <NewPost>
-            <button onClick={openNewPostModal}>
-              <FaNewspaper color="rgba(0,0,0,0.6)"/>
-              <span>Comece uma publicação</span>
-            </button>
-            <NewPostBackground id="newPostModalBackground">
-              <NewPostModal>
-                <FaTimes id="closeModal" onClick={closeNewPostModal}/>
-                <span>Crie uma publicação</span>
-                <hr/>
-                <form onSubmit={handleNewPost}>
-                  <input
-                    type="file"
-                    name="post-image"
-                    id="post-image"
-                    accept="image/*"
-                    onChange={onSelectFile} />
-                  <img className="preview-img hidden" alt="preview"></img>
-                  <ReactCrop
-                    className="react-crop"
-                    src={upImg}
-                    onImageLoaded={onLoad}
-                    crop={crop}
-                    onChange={c => setCrop(c)}
-                    onComplete={makeClientCrop}
-                  />
-                  <button className="cropImage hidden" onClick={previewImage}>Cortar</button>
-                  <label htmlFor="post-image" className="label-preview"><MdAddAPhoto color="rgba(0,0,0,0.6)"/></label>
-                  <textarea
-                    name="post-body"
-                    id="post-body"
-                    cols="30"
-                    rows="10"
-                    value={newPostContent}
-                    onChange={(e) => setNewPostContent(e.target.value)}
-                    placeholder="No que você está pensando?" />
-                  <hr/>
-                  <button type="submit">Publicar</button>
-                </form>
-              </NewPostModal>
-            </NewPostBackground>
-          </NewPost>
+          <NewPost />
           <FeedHistory>
             {
               posts.map(post => (
